resourceloader: Decouple FileContentsHasher from MW ObjectCache
Decouple from MW-specific ObjectCache class and stateful servie
wiring that lies behind it, which currently make it impossible
to test FileModule in a pure unit test. This is also in prep for
resourceloader being standalone.

Remove fallback to HashBagOStuff, previously used if a server
doesn't have php-apcu. This fallback didn't seem useful as it
would only retain fairly large amounts if data in memory, that
is actually not re-used as we generally don't hash the same file
multiple times, or at least not in significant quantities that it
outweighs the out-of-the-box risk in memory use for large RL
startup responses where many 10,000s of files get hashed.

Bug: T32956
Change-Id: I809951d34feb2dbd01b7ae0f4bd98dac7c3f6fe2