resourceloader: Remove support for $algo param in FileContentsHasher
This isn't used anywhere. Theoretical code using would gracefully
fallback to 'md4' given extra params are ignored by PHP.

In terms of future roadmap, I do expect that this algo may have to
change one day based on information that might come to light around
whether MD4 will remain sufficiently fast and/or sufficiently unique.

However, I do not expect that to be so significant in difference that
we'd have to remain some callers on MD4 whilst others adopt something
else. It seems fine that, if that time comes, to simply switch it
and thus consider the algo an implementation detail.

Bug: T32956
Change-Id: Ia9e1366b802ac18f439ce0c802189ede0a2c63f0