installer: Various tweaks for taint-check
Taint-check really doesn't like assigning things from $GLOBALS this way,
because it knows that stuff inside $GLOBALS can be tainted (e.g.
$GLOBALS['_GET'], $GLOBALS['_POST']), and since it cannot determine
which key is used in the assignment, it assumes that it's tainted.
As such, the new code patch generates the default values in a separate
method, whose return value is annotated as safe, so taintedness from
$GLOBALS doesn't get into $this->settings.

Also annotate getAcceptLanguage() as untainted, since the raw values
from the HTTP header are matched against a list of known codes.

Change-Id: Ifc44d3703d505962e3b07db7613b5bbf5e448320