skins: Change Skin::getPageClasses to avoid Title::getRootTitle
Title::getRootTitle will throw on invalid page titles. This can happen
when the page title contains a prefix that clashes with a registered
namespace or interwiki. For example when pages exist on a wiki, and
then the namespace or interwiki is registerered. These are meant to
be converted by maintenance/namespaceDupes.php.

The invalid page title could be viewed via curid= parameter.

In such situation, there is no way to re-create the Title object for
the given namespace and title. The only place where this happens during
a pageview is when the skin is trying to create a Title object for
the parent title (which is a shorter text of the otherwise mostly the
same title in the same namespace).

Workaround this by creating the string directly via the TitleFormatter
instead of creating the impossible Title object in the middle.

Bug: T290194
Change-Id: I946508ab7f95e391e0b71e1a3f0c81dcf12f73ee