Rename variables to disambiguate OutputPageParserOutput hook
It can be difficult to distinguish between ParserOutput and OutputPage
values, since both are often named $out or $output.  Use more
distinctive variable names to make it easier to read the code (and use
code search) for uses of the OutputPageParserOutput hook, where both
types are present in close proximity.

See T293860 for an instance of this confusion biting us in production,
and I386e99df01f889a80de26e861142bdb606aa649d for a related tidying-up
in Wikibase.

Change-Id: I0ccb7484eb3b2e2ae6a5a85c3e08d7d48b1e6146