jobqueue: Fix max no. of double redirect jobs in job queue
Introduce a constant to track the maximum number of double redirects
jobs in the job queue. Before we allow 10000 to spread in the code
base, we can just change the constant now and references of this value
will be changed.

Does this qualify to be a configuration or a regular constant will do?

Change-Id: I10c45a8c1cd5b3382edebf9a05f4b3f94f5327df