Reduce delete attempts on redirect table for move and page creation
Avoid delete of non existing rows which sets gap locks.

For PageUpdater::doCreate: When creating a new page it is known there is
no row in the redirect table, but a "DELETE FROM redirect WHERE page_id
= new page" is running within the open transaction and sets a gap lock
for "last page id in redirect" until unlimited. A redirect page creation
from another user now wants to insert and has to wait for release of the
lock, because it has a page id, which is in the locked gap. That can
result in timeouts for page creation.

For MovePage::moveToInternal: When a normal page is moved it is known
there is no row in the redirect table, but the page move is issue a
delete + gap lock and a parallel move with a page id in that range can
get a deadlock.

Bug: T120140
Change-Id: I5ddfdb690e0cc33b178b5fdcc01ffb6220911fa8