Provide button for revision deletion of selected files on file page
Use a button plus checkbox to allow selection of the file versions to
revision delete/suppress and remove link for only one file.

The underlying Special:RevisionDelete already support to work with more
than one file version. There is no change needed for that page or for
the permission checks needed for revision deletion.
Just building up the correct url (also possible by hand).

The header for the file history section is moved to the pager to allow
adding the form tag around it, which is needed to submit on the "change
visibility" button as done on Special:Log and action=history

Bug: T270750
Change-Id: I8b8359376edd4e5baf3bcfd755cb620ea3fc1a8a