Move page title to the beginning of page action interface messages
This was already changed in core, and reverted, in 2012
because it looked like the discussions are not conclusive:
Iad2c25fbbb673ad95781b29630193bc9de97b1be.

However, this has been the de-facto state in the English Wikipedia
since 2012. The community's explanation for this is that
it puts the title in the begining of the page title
and this is convenient for reading the title, especially
with small browser tabs. This explanation makes sense to me, too.

Furthermore, it was customized in the same manner in
the translations of these messages to many languages on
translatewiki:
* https://translatewiki.net/wiki/Special:Translations?message=difference-title&namespace=8
* https://translatewiki.net/wiki/Special:Translations?message=difference-title-multipage&namespace=8
* https://translatewiki.net/wiki/Special:Translations?message=history-title&namespace=8

So it looks like that's what many wiki users in
many languages actually want.

Change-Id: I59202fb96460fe2ecf5958674cf6a7bf1cc843f4