installer: Set "unknown" for default value of *_major_mime in MySQL
Follow up to Ic45dc1bce796a0 which added a schema change that mistakenly
removes the default value and given that it was done long time ago and
the main patch was removed since then, we can simply add a new patch for
broken cases.

Sqlite and Postgres don't need this change as they didn't get the patch
in the first place.

Also fixing the nullability of these fields.

Bug: T277354
Change-Id: Iced6fc6a7127a33cc171174270eba5cc552c871e