Settings: Prevent applying more settings after invoking `->finalize()`
When we're done applying settings and we call `finalize()`, any further
attempt to apply settings too late should throw.

This is a defensive mechanism for us to track the state of settings applied
in during setup process and will guide us to bring all settings to be applied
in the appropriate place so unexpected things don't happen.

Bug: T296684
Change-Id: Id58654c8265bc8310e1fe464ff5a480aad8a8666