Skin: Disable the new main page messages by default
Follows-up I36332565, some wikis do show a heading on the main page e.g.
nl.wiktionary.org, currently as it stands these wikis would see their
heading change to "Main page" in the event of the message not being
translated. To counter this, we can disable the message by default and
use its disabled state to determine whether to apply the customizations.

Bug: T255682
Change-Id: I257024446697ac6e4b8ae594dd7476a3d2b898d8