Allow editors to blank the heading, standardize heading class
Add a new skin template field which represents the heading element
and allow it to be blanked to result in it being hidden.

Skins previously had to provide the ID/class of the heading element
This led to Minerva using an inconsistent ID. Take the opportunity
to standardize this by adding an mw-first-heading class.

After this change we'll update Minerva and Vector to use this new
value.

Bug: T297019
Change-Id: Ife2fe5027519160ac4a130c6b460e946a2b834ff