Link to working Special:UserRights for users with '@' in their name
The value of $wgUserrightsInterwikiDelimiter ('@') is forbidden in new
names, however it is entirely possible that legacy accounts still use an
@ character. In those cases, linking to Special:UserRights/{username}
does not work, as it will interpret the @ as an interwiki delimiter.

Special:UserRights supports an alternative syntax using the user ID to
work around this issue. So we can check if the username contains the
delimiter, and if so, link via the user ID rather than by name.

I tested this by setting `$wgUserrightsInterwikiDelimiter = 'W';` and
looking at the link generated on [[User:MediaWiki default]]

While we're at it, fix the documentation for makeSpecialUrlSubpage() to
indicate that $subpage can be false.

Bug: T260222
Change-Id: Ib22ae8afdec147001577554fd662300df6cc7a11