Very long words in editcomments should wrap
Very long words in edit summaries on pages likes RC, watchlist and
diff pages would break out of their normal layout boxes. Apply
break-word to make sure that if a word cannot find on a line, it is
broken in the middle.

Very long words are common with vandalism but also in valid content
like for instance long urls.

Bug: T158725
Change-Id: I3906087e8b28ad71178f2b44baee4d495432e173