phpunit: Use submap assertion for LanguageIntegrationTest
LanguageIntegrationTest asserts strict equality for namespaces, but
other extensions which add to namespaces are not accounted for. This
seems to work OK when LanguageIntegrationTest is run alongside all other
PHPUnit tests but causes failures when run on its own (see console
output in T297078).

Asserting the submap should be sufficient for this test, AIUI.

Bug: T297079
Change-Id: I708c79f452a3eede1bae1422f27160a32db0e4dd