SECURITY: Do not apply autoblocks to untrusted XFF headers
X-Forwarded-For is not necessarily trustworthy and can specify multiple
IP addresses in a single header, all of which are checked for blocks.
When a user is autoblocked, the wiki will create an IP block
behind-the-scenes for that user without exposing the user's IP on-wiki.
However, spoofing XFF would let an attacker guess at the IPs of users
who have active autoblocks, since the block message includes the
username of the original block target.

We still want to apply autoblocks to XFF when it comes from a trusted
proxy server so that autoblocks still work on wikis with reverse proxies
in front of them. However, don't allow potentially-spoofed XFF through
to close up this privacy loophole.

Bug: T285159
Change-Id: I3d7d9ac57757be002f27db13f170ff9455a4e5b9