Move the link updates from moveUnsafe() to LinksUpdate
Since LinksUpdate was already run after a page move, it was just a
matter of propagating a move flag down to it, and resolving the
differences between the move case and the edit case.

This should be faster since it avoids writing to the links tables twice
on a page move.

Change-Id: Ifaa4051a736b01cda82fbd890325e0751ce041f8