Fix bottom margin for thumbs
Follow up to 7eb9972 where the ordering of margins for .tright and .tleft
and .thumb reversed, giving the latter higher precedence.

A consequence of this patch is that the margin is now applied to print
styling as well.  However, that seems consistent with all the other
thumbnail margins and really only affects "center" and "none", where
this bottom border isn't overwritten.

Change-Id: Idd9ba6adbaee3de8cc9a689314debaa0eb31f9d3