objectcache: handle object tokens in MediumSpecificBagOStuff::doCas()
Some subclasses, like APCUBagOStuff, use the whole cache value itself
as the CAS token. Previously, in such cases, an error would be thrown
when the key held a PHP object, and the comparison would fail if the
key held an array that contained an object.

Make RedisBagOStuff and HashBagOStuff return null CAS tokens if the
value is false, to match other subclasses.

Make WinCacheBagOStuff CAS tokens distinguish int and digit strings.

Change-Id: I797f036e53aae7a4fdcb8586fbd4ebdbd7e8bae3