SECURITY: Add recursion guard to Title::newMainPage()
If the on-wiki "mainpage" message is a bare local interwiki
(`localiw:`), MediaWiki will try to normalize it to the main page,
causing an infinite recursive loop. Once in this state, there's no way
to fix it short of editing the database (or applying this patch).

The easiest way out of this is to add a recursion guard to fall back to
the hardcoded "Main Page" default.

Bug: T297571
Change-Id: I91316ecbdbe7c4ad1b7594643aef29a8e1f8fb26