Remove unreachable parts of getTitleInvalidRegex() in PHP and JS
In a3a2744d03 (then r28968), the illegal title character regex was
expanded to block XML/HTML character references. However even back then
the two parts of the regex that aimed to block character references with
hashes in them were unreachable, because the fragment split happened one
step before illegal characters were checked. Any character reference
with a # in it would be split across the dbkey and fragment, and the
regex is only run against the dbkey.

This was noted in the JavaScript port of mediawiki.Title's tests, but
still left in. Those commented out test cases have now been removed as
they're technically valid titles with fragments.

Bug: T297578
Change-Id: I079f91a87fd7587ed70e65916fdd5efa01735efe