Add "resource" attribute to img tags
This is spec'd out at https://www.mediawiki.org/wiki/Specs/HTML#Media

It's also useful in the bug to determine when the link is pointing at
the resource, and hence MediaViewer should open.

Previously that was distinguished with .image class on the link but
that's now omitted in getDescLinkAttribs.

FIXME: Should the "resource" contain querystrings?  Maybe this needs to
be done on the Parsoid side as well.

Bug: T292657
Depends-On: Idb60e418f79dcb6a121de2a11e6e0ed0b31fd3ff
Change-Id: Ia94138383ebdbfc2feef75fdf651b969085a72b1