Allow EditPage's preload feature only on wikitext pages
This is a convenience feature that makes it possibly to
semi-automatically copy-paste the contents from another page when
creating a new one. The more I dig into the code, it's history and
potential use-cases the more I feel like this was never meant to be
used on non-wikitext pages.

* This feature makes it possible to trick users into making edits to
e.g. one of their personal .js subpages. I find this scary.

* "preload" and "preloadparams" are meant to work like a template
transclusion where "preload" is the name of the template. I don't
think this makes sense on non-wikitext pages.

* The feature can be used together with section=new. This doesn't
work anyway on non-wikitext pages. The only effect this patch will
have is when a non-existing page is created.

* The feature is meant to seed e.g. a new section on a talk page with
something the user can work with before it is submitted.

* But what's the point of semi-automatically seeding e.g. a .css or
.js subpage with the contents from another page? One situation I can
think of is when a new user gets the recommendation to copy-paste
another users skin modifications. It might be convenient to let
the user click a link that utilizes the preload feature instead of
manually copy-pasting code. But this comes with so many problems (e.g.
the user doesn't have a chance to understand what's going on) I don't
think it's worth it.

Bug: T297725
Change-Id: Iee2d9196854427501bf36659eace81a80a66dd26