Remove prefix search from Special:ListFiles/NewFiles
This affects only wikis without miser mode.
The database LOWER function may not work,
if the database field has the wrong collation.
The use of Title::newFromText may apply normalization which are not
useful for prefix/like searches as done here.

This should be done by a search backend, which can do that with better
performance.

Bug: T298722
Change-Id: I553b25794962f71d7588d735ced88a2dc3812093