ResourceLoader: Set module state "error" if request fails on network
If a script fails to load at the network level, e.g. if your WiFi
is on when the page first renders but then is turned off when you
later use an interactive feature that needs additional scripts,
then the additional modules would be marked neither as "ready" nor
as "error". Instead, the remained stuck as "loading" because the
server response would call neither the mw.loader.implement callback,
nor mw.loader.state(), as there is no server response, not even an
"error" response.

In addScript(), add a setAndPropagate() call in the native
`script.onerror` event handler to detect this scenario and let it
mark the affected modules as state "error".

This means that UI code that uses `mw.loader.using()` to request
modules lazily, we now correctly reject the returned Promise, which
means any fail/catch handlers the UI might have for handling errors
like "Syntax error" or "Missing dependencies" will now also run for
this use case.

Bug: T68598
Change-Id: I825b09273c73a1ddfc080af8044541a456679fb0