Add Nheengatu (yrl) to Names.php
The autonym "ñe'engatú", which appears in the current configuration
of translatewiki and language-names in ULS, also appears in
Ethnologue, but it is probably incorrect, as it doesn't appear
almost anywhere else.

There are several autonym variants, and the one that appears
the most consistent with the actual translations in translatewiki
is "Nhẽẽgatú".

Bug: T298385
Change-Id: I4d794dd50475c0ef17cbd906801d5a53d8446778