Add per-table straight join option
Add support for MySQL's per-table syntax for straight joins, which is
a join type analogous to LEFT JOIN.

SelectQueryBuilder::straightJoin() previously enabled the whole-query
straight join option, but it makes more sense for it to work analogously
with leftJoin(). So I renamed the existing method to
straightJoinOption(). No callers found in CodeSearch.

The motivation is to allow us to prevent reordering of the change_tag
join without affecting the rest of the recentchanges query.

Bug: T298225
Change-Id: I2bfcffc2bf9a64f23afdea0dfaccf89938896e62