Remove unused EditPage::$mPreloadContent property
Unused:
https://codesearch.wmcloud.org/search/?q=%3E(mPreload%7CsetPreloaded)(Text%7CContent)&files=%5C.php%24

There was a public setter `setPreloadedContent()` that was removed in
2020 in patch Id47421e, tracked via T252907.

The fact that the property is marked as public is most certainly not
intentional. It was introduced when `private` didn't exist. Not only
that. It was dynamically created and never declared. This was fixed
in 2011, see
https://phabricator.wikimedia.org/rMWaa9bc3ba41273f911bad6a3ac754c98e1dde229a
The assumption was made that the property was meant to be public, but
I think this was a mistake.

Bug: T297725
Change-Id: Ic6a12d706c823c31b8e8a736f078977ec64be973