htmlform: Avoid executing cond-state logic twice on Special:Preferences
The 'htmlform.enhance' hook would be fired on the document root and every tab element on Special:Preferences. All elements will be infused by this script in the first run on the document root, which aginst the 'autoinfuse-lazy' design of pref-tabs.

When most preferences are set as global, the script takes a relatively long time to finish.

Change-Id: Ie066516768290c254255d5604e98bcae6a78a545