Add support for namespace notices
There are use-cases for namespace-specific sitenotice functionality, for instance on the draft namespace.

Content of MediaWiki:Namespacenotice-N will be shown along with the normal sitenotice/anonnotice on pages in namespace number N. Like the regular sitenotice, this is cached on a per-namespace level, so it does NOT support page-specific notices using {{PAGENAME}} or the like.

Bug: T6469
Change-Id: I091ed3791681e4c7044e79cdb14ebdd123edc15c