LoginSignupSpecialPage: Add hide-if statement conditionally and use field name
Field name which is used as the key in html descriptor is required.
After I58f9df384df8ecc5ebae8cac68ec2251351bc984, values of fields that are supposed to be disabled would be loaded from default.
Nothing is broken here since the operator is '===', just for my follow-up patch (I9a42417a6161f42181badd8cdbec81ba85dc62f6) which will throw exceptions when using a mismatched field name.
The 'name' param is kept to make sure everything work, and passing the correct element id to the client-side is the new feature in the follow-up change, so this should be merged right before it.

Change-Id: I5ed953ec104fc4739453d1da756a69efe743abd2