Fix PHP 8.0 failure of WikiExporterFactoryTest
In PHP 7, "some unlikely string 2" was a valid value for the $text
parameter to XmlDumpWriter::__construct(), because non-strict comparison
was used, so "some unlikely string 2" was coerced to integer 0 which is
equal to WRITE_CONTENT. In PHP 8.0, non-strict comparison has changed such
that "some unlikely string 2" != 0.

Pass strict=true to in_array() so that the test fails on both versions.
Please review for potential production impact.

Fix the test so that it passes after this change.

Bug: T283275
Change-Id: I5a84fb70db9d02ea10f87299eb41f80ba8fc787c