Untie Wikimedia message boxes from on-wiki messageboxes
Many communities for historic reasons use "messagebox" to style
certain templates. Going forward we aim to make sure all internally
produced CSS classes are prefixed with "mw-".

A user notice has been run notifying editors that this is potentially
a breaking change with directions on how to update their CSS in
preparation for this change.

Bug: T270796
Depends-On: Id258e328d1691f7bd31d2b3c57d64d5434706561
Change-Id: I2df24431ec72641b63fbe2fc929025a53459c06c