SettingsBuilder: fix merging into globals
GlobalConfigBuilder needs to use the appropriate merge strategy to
combine new values with the value already rpesent in a global variable.
For this, GlobalConfigBuilder needs to actually look at the current
value of the global variable. Also, we need to define merge strategies
for all settings that need them (at least the ones used directly by
ExtensionProcessor).

Change-Id: I834d4b3506bdd8d416e5eb6e03fb4dd6b60b6e05