Linker::getRollbackEditCount confused on revisions with same timestamp
When revisions have the same timestamp the rollback could be confused
about the order of both revisions.
Add additional order by rev_id as used in
RevisionStore::getRelativeRevision to fix this situation

mysql is ordering by the primary key in case of same values of an index,
using the same order, but postgres fails in this situation.

Bug: T195807
Bug: T218400
Change-Id: Id40f8351992de0fb747f09821de4e8463c8986d2