Update category counts on purge
We only increment/decrement category counts when a page is added or
removed,
which means once they get out of sync, they'll stay that way forever or
until a maintenance script is run.

Now using action=purge (via index.php or Action API) will also cause the
category counts to be refreshed if the category has less than 5000
items,
as suggested on the linked bug.

Category::refreshCounts() already has locking logic so if someone is
already
purging a category, a second user won't be able to as well.

Bug: T85696
Change-Id: Ib21d8d63445db21591cd764d7d578191943c95f9