Fix signature of DatabasePostgres::buildGroupConcatField
The method in the postgres layer has an extra argument in the middle,
making the last parameter unused and the join condition is not passed to
build the subquery.
For usage of that function with a join the join condition are missing
and making the subquery returning a cartesian product, in case of change
tags with all existing tags in the wiki.

The code is there since the addition of the function in a974025
It is first used in a in a5500e7 for the change tags

Bug: T195807
Bug: T256401
Change-Id: I4e73e37d176a6368e65aeb21e5dbed14edd24f19