Add timestamp to thumbnail URLs on file pages
When overwriting a file, the browser will sometimes display a cached old
version of the file instead of the new version. This is because the URLs
for the current version of a file (and thumbnails) are not versioned.

This patch adds the timestamp as a query parameter to the end of the
file URLs. Only the src and srcset URLs will have the timestamp, the
links to the original and thumbnails are not versioned. The current
version in the file history also gets the timestamp. Previous versions
already have a timestamp in the URL.

This timestamp is only used for client-side cache busting, it is not
interpreted server-side. On WMF sites it will be stripped out by the
caching layer anyway.

Bug: T38380
Change-Id: Ia63bd96a02d1aa36265742c4307f5af2e675b3ec